13.11: update Linaro firmware overlay version
Updated the Linaro firmware overlay ZIP to provide the latest versions
of UEFI.

The latest version includes the feature to skip the initrd if it is not
present in the /boot partition.

OpenEmbedded images do not ship with an initrd, whereas Android images
do, so 1 config would no longer with both without this feature.

Signed-off-by: Ryan Harkin <ryan.harkin@linaro.org>